pyproject.toml

[tool.poetry]
name = "hastie"
version = "0.9.6"
description = "A static site generator"
authors = ["Marcus Kazmierczak <marcus@mkaz.com>"]
license = "MIT"
readme = "README.md"
keywords = ["static site generator", "static site", "blog"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: Content Management System",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup :: Markdown",
]

[tool.poetry.dependencies]
python = ">=3.10"
jinja2 = "^3.1.2"
markdown = "^3.4.1"
pygments = "^2.14.0"
python-frontmatter = "^1.0.0"
toml = "^0.10.2"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
pytest = "^7.2.1"
mypy = "^1.0.1"

[tool.poetry.scripts]
hastie = 'hastie.main:main'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
pythonpath = "hastie"
testpaths = [
    "tests",
]

[tool.ruff]
target-version = "py310"

[tool.ruff.per-file-ignores]
"tests/test_content.py" = ["E402"]
"tests/test_hfs.py" = ["E402"]

[tool.mypy]

[[tool.mypy.overrides]]
module = [
    "markdown",
    "toml"
]
ignore_missing_imports = true
